requirements.txt
Django==4.1.7
python-decouple==3.7

djangorestframework~=3.14.0